+++ changelog.md
@@ -1,3 +1,7 @@
+# 1.0.4
+- Text object is now removed from the level during the editing (unless in debug mode)
+- Change save object coordinates because of issue in another mod
+
 # 1.0.3
 - Add read-only access to the text object data on `PlayLayer` (requested by Balan4ik)